make favoriteToolsVM reusable for all favorites view
