+++ godtools/App/Features/Tools/SwiftUI/Favorites/AllFavoriteToolsView/AllFavoriteToolsViewModel.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class AllFavoriteToolsViewModel: ToolCardProvider {
+class AllFavoriteToolsViewModel: BaseFavoriteToolsViewModel {
     
 }
