+++ godtools/App/Features/Tools/SwiftUI/Favorites/FavoriteToolsView/FavoriteToolsViewModel.swift
@@ -13,60 +13,29 @@ protocol FavoriteToolsViewModelDelegate: ToolCardDelegate {
     func viewAllFavoriteToolsButtonTapped()
 }
 
-class FavoriteToolsViewModel: ToolCardProvider {
+class FavoriteToolsViewModel: BaseFavoriteToolsViewModel {
  
     // MARK: - Properties
-    
-    private let dataDownloader: InitialDataDownloader
-    private let deviceAttachmentBanners: DeviceAttachmentBanners
-    private let favoritedResourcesCache: FavoritedResourcesCache
-    private let languageSettingsService: LanguageSettingsService
-    private let localizationServices: LocalizationServices
-    private weak var delegate: FavoriteToolsViewModelDelegate?
+    private var favoriteToolsViewModelDelegate: FavoriteToolsViewModelDelegate? {
+        return delegate as? FavoriteToolsViewModelDelegate
+    }
     
     // MARK: - Published
     
-    @Published var sectionTitle: String = ""
     @Published var viewAllButtonText: String = ""
     
     // MARK: - Init
     
     init(dataDownloader: InitialDataDownloader, deviceAttachmentBanners: DeviceAttachmentBanners, favoritedResourcesCache: FavoritedResourcesCache, languageSettingsService: LanguageSettingsService, localizationServices: LocalizationServices, delegate: FavoriteToolsViewModelDelegate?) {
-        self.dataDownloader = dataDownloader
-        self.deviceAttachmentBanners = deviceAttachmentBanners
-        self.favoritedResourcesCache = favoritedResourcesCache
-        self.languageSettingsService = languageSettingsService
-        self.localizationServices = localizationServices
-        self.delegate = delegate
         
-        super.init()
-        
-        setupBinding()
-        setText()
+        super.init(cardType: .favorites, dataDownloader: dataDownloader, deviceAttachmentBanners: deviceAttachmentBanners, favoritedResourcesCache: favoritedResourcesCache, languageSettingsService: languageSettingsService, localizationServices: localizationServices, delegate: delegate)
     }
     
-    deinit {
-        dataDownloader.cachedResourcesAvailable.removeObserver(self)
-        dataDownloader.resourcesUpdatedFromRemoteDatabase.removeObserver(self)
-        favoritedResourcesCache.resourceFavorited.removeObserver(self)
-        favoritedResourcesCache.resourceUnfavorited.removeObserver(self)
-        favoritedResourcesCache.resourceSorted.removeObserver(self)
-        languageSettingsService.primaryLanguage.removeObserver(self)
-    }
-    
-    // MARK: - Overrides
-    
-    override func cardViewModel(for tool: ResourceModel) -> BaseToolCardViewModel {
-        return ToolCardViewModel(
-            cardType: .favorites,
-            resource: tool,
-            dataDownloader: dataDownloader,
-            deviceAttachmentBanners: deviceAttachmentBanners,
-            favoritedResourcesCache: favoritedResourcesCache,
-            languageSettingsService: languageSettingsService,
-            localizationServices: localizationServices,
-            delegate: delegate
-        )
+    override func setText() {
+        let languageBundle = localizationServices.bundleLoader.bundleForPrimaryLanguageOrFallback(in: languageSettingsService)
+        viewAllButtonText = localizationServices.stringForBundle(bundle: languageBundle, key: "favorites.favoriteTools.viewAll") + " >"
+        
+        super.setText()
     }
 }
 
@@ -74,137 +43,6 @@ class FavoriteToolsViewModel: ToolCardProvider {
 
 extension FavoriteToolsViewModel {
     func viewAllButtonTapped() {
-        delegate?.viewAllFavoriteToolsButtonTapped()
-    }
-}
-
-// MARK: - Private
-
-extension FavoriteToolsViewModel {
-    private func setupBinding() {
-        
-        dataDownloader.cachedResourcesAvailable.addObserver(self) { [weak self] (cachedResourcesAvailable: Bool) in
-            DispatchQueue.main.async { [weak self] in
-                if !cachedResourcesAvailable {
-                    // TODO: - add loading & find tools views
-//                    self?.isLoading.accept(value: true)
-//                    self?.hidesFindToolsView.accept(value: true)
-                }
-                else {
-                    // TODO: - add loading view
-//                    self?.isLoading.accept(value: false)
-                    self?.reloadFavoritedResourcesFromCache()
-                }
-            }
-        }
-        
-        dataDownloader.resourcesUpdatedFromRemoteDatabase.addObserver(self) { [weak self] (error: InitialDataDownloaderError?) in
-            DispatchQueue.main.async { [weak self] in
-                // TODO: -
-//                self?.didEndRefreshing.accept()
-                if error == nil {
-                    self?.reloadFavoritedResourcesFromCache()
-                }
-            }
-        }
-        
-        favoritedResourcesCache.resourceFavorited.addObserver(self) { [weak self] (resourceId: String) in
-            DispatchQueue.main.async { [weak self] in
-                self?.addFavoritedResource(resourceId: resourceId)
-            }
-        }
-        
-        favoritedResourcesCache.resourceUnfavorited.addObserver(self) { [weak self] (resourceId: String) in
-            DispatchQueue.main.async { [weak self] in
-                self?.removeFavoritedResource(resourceIds: [resourceId])
-            }
-        }
-        
-        favoritedResourcesCache.resourceSorted.addObserver(self) { [weak self] (resourceId: String) in
-            DispatchQueue.main.async { [weak self] in
-                self?.reloadFavoritedResourcesFromCache()
-            }
-        }
-        
-        languageSettingsService.primaryLanguage.addObserver(self) { [weak self] (primaryLanguage: LanguageModel?) in
-            DispatchQueue.main.async { [weak self] in
-                self?.setText()
-            }
-        }
-    }
-    
-    private func addTool(tool: ResourceModel) {
-        
-        var updatedToolsList: [ResourceModel] = tools
-        
-        guard !updatedToolsList.contains(tool) else {
-            return
-        }
-        
-        updatedToolsList.insert(tool, at: 0)
-        tools = updatedToolsList
-    }
-
-    private func removeTools(toolIdsToRemove: [String]) {
-        let toolsToRemove: [ResourceModel] = tools.filter({toolIdsToRemove.contains($0.id)})
-        removeTools(toolsToRemove: toolsToRemove)
-    }
-    
-    func removeTools(toolsToRemove: [ResourceModel]) {
-        
-        guard !tools.isEmpty && !toolsToRemove.isEmpty else {
-            return
-        }
-        
-        var updatedToolsList: [ResourceModel] = tools
-        var removedIndexPaths: [IndexPath] = Array()
-        
-        for toolToRemove in toolsToRemove {
-            if let index = updatedToolsList.firstIndex(of: toolToRemove) {
-                removedIndexPaths.append(IndexPath(row: index, section: 0))
-                updatedToolsList.remove(at: index)
-            }
-        }
-        
-        withAnimation {
-            tools = updatedToolsList
-        }
-    }
-    
-    private func addFavoritedResource(resourceId: String) {
-        if let tool = dataDownloader.resourcesCache.getResource(id: resourceId) {
-            addTool(tool: tool)
-            // TODO: -
-//            hidesFindToolsView.accept(value: !tools.value.isEmpty)
-        }
-    }
-    
-    private func removeFavoritedResource(resourceIds: [String]) {
-        removeTools(toolIdsToRemove: resourceIds)
-        // TODO: -
-//        hidesFindToolsView.accept(value: !tools.value.isEmpty)
-    }
-    
-    private func reloadFavoritedResourcesFromCache() {
-        
-        let sortedFavoritedResources: [FavoritedResourceModel] = favoritedResourcesCache.getSortedFavoritedResources()
-        let sortedFavoritedResourcesIds: [String] = sortedFavoritedResources.map({$0.resourceId})
-        
-        let resources: [ResourceModel] = dataDownloader.resourcesCache.getResources(resourceIds: sortedFavoritedResourcesIds)
-        
-        let filteredResources: [ResourceModel] = resources.filter({
-            return !$0.isHidden
-        })
-        
-        tools = filteredResources
-        // TODO: - fix this
-//        hidesFindToolsView.accept(value: !filteredResources.isEmpty)
-//        isLoading.accept(value: false)
-    }
-    
-    private func setText() {
-        let languageBundle = localizationServices.bundleLoader.bundleForPrimaryLanguageOrFallback(in: languageSettingsService)
-        sectionTitle = localizationServices.stringForBundle(bundle: languageBundle, key: "favorites.favoriteTools.title")
-        viewAllButtonText = localizationServices.stringForBundle(bundle: languageBundle, key: "favorites.favoriteTools.viewAll") + " >"
+        favoriteToolsViewModelDelegate?.viewAllFavoriteToolsButtonTapped()
     }
 }
